+++ tests/hash/sha224.c
@@ -2031,7 +2031,7 @@ static struct QC_SHA224_TestVector vectors[] = {
 
 int main()
 {
-    int i, j;
+    size_t i, j;
 
     for (i = 0; i < sizeof(vectors) / sizeof(vectors[0]); i++)
     {
@@ -2055,7 +2055,7 @@ int main()
                 printf("%02x", vectors[i].sha224[j]);
             }
 
-            printf(" in test vector %d\n", i);
+            printf(" in test vector %zu\n", i);
             return 1;
         }
     }
@@ -2082,7 +2082,7 @@ int main()
                 printf("%02x", sha224_test_sha224[i][j]);
             }
 
-            printf(" in test vector %d\n", i);
+            printf(" in test vector %zu\n", i);
             return 1;
         }
     }
