+++ tests/hash/sha256.c
@@ -2037,7 +2037,7 @@ static struct QC_SHA256_TestVector vectors[] = {
 
 int main()
 {
-    int i, j;
+    size_t i, j;
 
     for (i = 0; i < sizeof(vectors) / sizeof(vectors[0]); i++)
     {
@@ -2061,7 +2061,7 @@ int main()
                 printf("%02x", vectors[i].sha256[j]);
             }
 
-            printf(" in test vector %d\n", i);
+            printf(" in test vector %zu\n", i);
             return 1;
         }
     }
@@ -2088,7 +2088,7 @@ int main()
                 printf("%02x", sha256_test_sha256[i][j]);
             }
 
-            printf(" in test vector %d\n", i);
+            printf(" in test vector %zu\n", i);
             return 1;
         }
     }
